+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased](https://github.com/elastic/package-registry/compare/v1.37.0...main)
+## [v1.38.0](https://github.com/elastic/package-registry/compare/v1.37.0...v1.38.0)
 
 ### Breaking changes
 
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Require local indexers to include signatures for packages. [#1646](https://github.com/elastic/package-registry/pull/1646)
 
 ### Bugfixes
+
+* Update Go runtime to 1.26.2 [#1675](https://github.com/elastic/package-registry/pull/1675)
 
 ### Added